Move tests to StaticProxyTests
