--- src/NHibernate.Test/NHSpecificTest/GH2619Proxy/Mappings.hbm.xml
@@ -1,13 +1,0 @@
-﻿<?xml version="1.0" encoding="utf-8" ?>
-<hibernate-mapping xmlns="urn:nhibernate-mapping-2.2"
-           namespace="NHibernate.Test.NHSpecificTest.GH2619Proxy"
-           assembly="NHibernate.Test"
->
-  <class name="ClassWithGenericNonVirtualMethod" table="CWGNVM" >
-    <id name="Id">
-      <generator class="increment"/>
-    </id>
-    <property name="Name"/>
-  </class>
-
-</hibernate-mapping>
